Make list items reuseable
